Update batchnorm (#25)
* Generic allreduce interface

- AllreduceMPICUDA -> cudaMemcpy and MPI_Allreduce
- AllreduceAlNCCL -> Use Al::NCCLBackend

Tested with test_allreduce_cuda.cu.
Used in in batchnorm.

* Test driver updates

- Add an option to use double-precision
- Disable bitwise validation in batchnorm
- Disable testing of multiple numbers of dimensions
- Add BN implementation option
- Fix paths

* Refactor batchnorm

* Measure allreduce times in batchnorm

* Batchnorm aggregation optimization

* Optionally skips allreduces of batchnorm weights

* Update CMake file for the allreduce headers

* Bug fix

* NVSHMEM-based allreduce naive implementation

* Simple recursive-doubling allreduce

* Recursive doubling with a single kernel

* Recursive doubling with separate buffers

* Allreduce with NVSHMEM's sum_to_all

* Block-wise allreduce

Use one thread block per element. Intended for use in batchnorm.

* Fused single-kernel forward batchnorm

* Bug fix

* Optimize batchnorm kernels

The existing implementations are not very efficient since they are
designed to be generic with respect to halo usage, which makes offset
calculation a bit complicated. These optimized versions are only used
for certain cases where it can be simplified.

* Batchnorm reduction with CUB

The improvement is very minor when tested with 128^3 tensors with 4
channels. The same optimization was also applied to the fused version,
but reverted as it resulted in slightly lower performance.