Patches for DistConv ReLU (#89)
* Correct copy-paste error in function argument

* Address in-place-ness in relu

There should also be some component of this in
distconv/dnn_backend/relu.hpp. When "output"=="input", one of "m_output_d"
or "m_input_d" will (likely) be the incorrect thing to pass (strides may
differ, for instance). The fix implemented here is in the
backend. While avoids this error with LBANN *and* potentially avoids
packing the same buffer twice, it's not clean-cut. It assumes that
"output" is "correct" and matches "m_output_d" while "input" may not
match "m_input_d". This matches the known use-case in LBANN.