Do not store a pointer from `RawMemory` in `StridedMemory`.
This avoids any issues where `RawMemory` changes the backing memory.
Instead we store a numerical offset and always use the `data` method
of `RawMemory` to ensure up-to-date pointers. This slightly
complicates the use of externally-managed memory.

This is to future-proof for cases involving, e.g., data stashing for
out-of-core algorithms.