Local tensor copy support (#123)
* Add check for whether our GPU is integrated (like an APU).

This currently is always false on CUDA and HIP/ROCm platforms. On the
latter, this is due to a bug in HIP's device properties.

* Clean up Tensor::resize.

Also don't reallocate if the size doesn't change.

* Add raw buffer copy, with device support.

* Support manually specifying the strides for `StridedMemory`.

Changes how buffer sizes are computed, to properly account for the
full extent defined by the strides. This fixes some corner cases where
we might calculate the wrong buffer size.

Added a `StridedMemory::size` method to return the actual buffer size;
this is mainly for testing use.

* Support resizing tensors and changing their strides.

* Initial support for copying local tensors.

* Update include/h2/tensor/copy.hpp

Co-authored-by: Tom Benson <benson31@llnl.gov>

---------

Co-authored-by: Tom Benson <benson31@llnl.gov>