Remove catch2 test cases from CTest (#142)
The old version required the SeqCatchTests binary to be runable at
build time, which was not always the case (the current code still
isn't cross-comile-safe, but we don't cross-compile ever). Moreover,
the failures we saw from this were often false positives (e.g.,
LD_LIBRARY_PATH was incomplete during the build phase but wouldn't be
under normal circumstances).

The new version just integrates the entire SeqCatchTests driver with
CTest. This means it shows as a single CTest test, but it doesn't have
to run at build time. The test will generate "the usual" stdout/stderr
for integration with CTest's default logging and it will output an XML
in JUnit format to make CI integration easier. The XML output will be
written to <build_dir>/test/unit_test/seq.xml.